+++ src/functions/receive/index.js
@@ -1,5 +1,6 @@
 import config from '../../config';
 import Facebook from '../../lib/facebook';
+import util from '../../lib/util';
 
   const response = {
     statusCode: 200,
@@ -39,15 +40,29 @@ import Facebook from '../../lib/facebook';
     } else if (event.httpMethod === 'POST') {
       console.log('POST request, indicates user input');
 
-      const facebook = new Facebook(config);
-
+      // Respond immediately to webhook
       callback(null, response);
-      const payload = JSON.parse(event.body);
-      facebook.sendReply(payload.entry[0].messaging[0])
-        .then((data) => callback(null, response))
-        .catch((err) => {
-          console.log('Error in request: ' + err.message);
-          callback(null, errorResponse);
-        });
+
+      // Verify signature
+      const hash = util.sha1(config.FACEBOOK_APP_SECRET, event.body);
+      const signed = util.compareSignatures(hash,
+        event.headers['X-Hub-Signature']);
+
+      if (signed === true) {
+        console.log('Request signature verified');
+        // Create bot instance
+        const facebook = new Facebook(config);
+        const payload = JSON.parse(event.body);
+        console.log(JSON.stringify(event));
+        facebook.sendReply(payload.entry[0].messaging[0])
+          .then((data) => callback(null, response))
+          .catch((err) => {
+            console.log('Error in request: ' + err.message);
+            callback(null, errorResponse);
+          });
+        } else {
+          console.log('Request signature did not match');
+          callback(null, 403, {});
+        }
     }
   };