+++ src/lib/facebook.js
@@ -148,6 +148,7 @@ export default class Facebook {
   **/
   sendThanks(body) {
     return new Promise((resolve, reject) => {
+      console.log(body);
       if (body.instanceRegionCode === 'null') {
         // catch reports outside the reporting area and reply a default
         body.instanceRegionCode = this.config.DEFAULT_INSTANCE_REGION_CODE;