+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # CHANGELOG
 
 ### v0.0.3
-
 Deployment version:
 - working bot with buttons
 - support for null areas
 - improved thanks message
+
+### v0.0.4
+- verify Facebook signatures
+- added India deployment
+- update bot-core package
+- improved test coverage